Componentizando a responsividade da navbar.
