+++ src/components/Header.jsx
@@ -1,19 +1,36 @@
 import styles from "./Header.module.css"
+import { useWindowHeight } from '@react-hook/window-size'
+import { useEffect, useState } from "react";
 
 import BloodyCastLogo from "../assets/logo.png"
-import { MenuElements } from "./MenuElements";
-import { SearchBar } from "./SearchBar";
-import { useState } from "react";
+import { NavbarDesktop } from "./NavbarDesktop";
+import { NavbarMobile } from "./NavbarMobile";
 
 export function Header() {
-    const [showMenu, setShowMenu ] = useState(false) 
-    const menuOptions = [
-        {optionName: "Home", optionUrl: "/"}, {optionName: "PodCast", optionUrl: "/PodCast"}, {optionName: "Cinema", optionUrl: "#"},
-        {optionName: "Livros", optionUrl: "#"}, {optionName: "Musícas", optionUrl: "#"}, {optionName: "Jogos", optionUrl: "#"}
-    ];
+    const [screenWidth, setScreenWidth] = useState(window.innerWidth);
+
+    const addScreenWidth = () => {
+        setScreenWidth(window.innerWidth);
+    }
     
-    const toggleMenu = () => {
-        setShowMenu(!showMenu)
+    useEffect(() => {
+        window.addEventListener('resize', addScreenWidth);
+
+        return() => {
+            window.removeEventListener('resize', addScreenWidth)
+        }
+    }, [screenWidth]);
+
+    const ScreenRender = () => {
+        if (screenWidth < 1024) {
+            return (
+                <NavbarMobile />
+            )
+        } else {
+            return (
+                 <NavbarDesktop />
+            )
+        }
     }
 
     return (
@@ -22,21 +39,7 @@ export function Header() {
                 <img className={styles.bloodyLogo} src={BloodyCastLogo} />
             </div>
 
-            <nav className={`${styles.menuSandwich} ${ showMenu ? styles.show : '' }`}>
-                {menuOptions.map(function(data) {
-                    return (
-                        <MenuElements name={data.optionName} screenUrl={data.optionUrl} />
-                    )
-                })}
-
-                <SearchBar />
-            </nav>
-
-            <div className={styles.menuButton} onClick={toggleMenu}>
-                <span className={styles.linha}></span>
-                <span className={styles.linha}></span>
-                <span className={styles.linha}></span>
-            </div>
+            <ScreenRender />
         </header>
     )
 }
