+++ src/components/Header.module.css
@@ -12,47 +12,7 @@
     margin-left: 25px;
 }
 
-@media (max-width: 1023px) {
-    .menuButton {
-        display: flex;
-        justify-content: space-evenly;
-        align-items: center;
-        flex-direction: column;
-    
-        width: 30px;
-        height: 30px;
-        right: 20px;
-        cursor: pointer;
-    }
-    
-    .linha {
-        display: block;
-        width: 100%;
-        height: 3px;
-        border-radius: 5px;
-        background-color: var(--white);
-    }
-
-    .menuSandwich {
-        display: none;
-    }
-
-    .show {
-        display: flex;
-        flex-direction: column;
-        gap: 1rem;
-
-        width: 80%;
-        height: auto;
-        position: absolute;
-        top: 125px;
-        right: 0;
-        background-color: var(--black);
-        padding: 20px;
-    }
-}
-
-@media (min-width: 1024px) {
+@media (min-width: 1024) {
     .container {
         display: flex;
         justify-content: space-between;
@@ -66,13 +26,6 @@
         height: 6.4rem;
         margin-left: 50px;
     }
-    
-    .nav {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        gap: 20px;
-    }    
 }
 
 @media (min-width: 1280px) {
@@ -89,13 +42,6 @@
         height: 6.8rem;
         margin-left: 50px;
     }
-    
-    .nav {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        gap: 25px;
-    }    
 }
 
 
@@ -113,13 +59,6 @@
         height: 7rem;
         margin-left: 50px;
     }
-    
-    .nav {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        gap: 30px;
-    }    
 }
