+++ src/components/MenuElements.module.css
@@ -1,7 +1,6 @@
 .activeContainer {
     background-color: var(--yellow-200);
     padding: 5px 10px;
-    border-radius: 8px;
 }
 
 .activeContainer a {
