+++ src/components/NavbarDesktop/NavbarDesktop.module.css
@@ -1,0 +1,25 @@
+.menuContainer {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    gap: 50px;
+}
+
+@media (min-width: 1280px) { 
+    .menuContainer {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        gap: 25px;
+    }    
+}
+
+
+@media (min-width: 1440px) {
+    .menuContainer {
+        display: flex;
+        flex-direction: row;
+        align-items: center;
+        gap: 30px;
+    }    
+}
