+++ src/components/NavbarMobile/NavbarMobile.module.css
@@ -1,0 +1,37 @@
+.menuButton {
+    display: flex;
+    justify-content: space-evenly;
+    align-items: center;
+    flex-direction: column;
+
+    width: 30px;
+    height: 30px;
+    margin-right: 25px;
+    cursor: pointer;
+}
+
+.linha {
+    display: block;
+    width: 100%;
+    height: 3px;
+    border-radius: 5px;
+    background-color: var(--white);
+}
+
+.menuSandwich {
+    display: none;
+}
+
+.show {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+
+    width: 80%;
+    height: auto;
+    position: absolute;
+    top: 125px;
+    right: 0;
+    background-color: var(--black);
+    padding: 20px;
+}
